qa_bot_project/requirements.txt
django==3.2.10
djangorestframework==3.12.4
langchain==0.0.184
openai==0.27.0
faiss-cpu==1.7.2
pypdf==3.8.1
tiktoken
